pyproject.toml
[tool.poetry]
name = "breslin-salesloft-eval"
version = "0.1.0"
description = ""
authors = ["TJ Breslin <tj@tjbreslin.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.1.2"
djangorestframework = "^3.12.1"
whitenoise = "^5.2.0"
gunicorn = "^20.0.4"
python-environ = "^0.4.54"
django-heroku = "^0.3.1"
requests = "^2.24.0"
PyHamcrest = "^2.0.2"
fuzzywuzzy = "^0.18.0"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
